% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{absolutizePath}
\alias{absolutizePath}
\title{Absolutize a Filepath}
\usage{
absolutizePath(x, dir = getwd(), winslash = "/", ...)
}
\arguments{
\item{x}{length one character: a file path}

\item{dir}{a reference directory}

\item{winslash}{path separator on windows, passed to \code{\link{normalizePath}}}

\item{...}{ignored arguments}
}
\description{
Absolutizes a filepath.  Somewhat the opposite of \code{\link{relativizePath}}.
}
\details{
x and dir are first normalized, then x is expressed relative to dir.
If x and dir are on different drives (i.e. C:/ D:/)
x is returned as an absolute path.
}
\keyword{internal}
