% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{as_calibrated.data.frame}
\alias{as_calibrated.data.frame}
\title{Calibrate Data Frame Column Widths}
\usage{
\method{as_calibrated}{data.frame}(x, width, ...)
}
\arguments{
\item{x}{A data frame.}

\item{width}{Character or numeric vector of column widths. Values are
recycled as necessary.}

\item{...}{Currently unused.}
}
\value{
A data frame with class \code{calibrated} and column-level
  \code{width} attributes.
}
\description{
Assign column-level \code{width} attributes to a data frame and return
a calibrated data frame for downstream table rendering. Widths are
recycled as necessary; a warning is issued when the number of supplied
widths differs from the number of columns.
}
\seealso{
\code{\link{as_calibrated}}, \code{\link[kableExtra]{column_spec}}

Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl}()},
\code{\link{kbl.calibrated}()},
\code{\link{kbl.default}()},
\code{\link{kbl.gridtable}()},
\code{\link{print.gridtable}()}
}
\concept{gridtable}
