% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{as_calibrated}
\alias{as_calibrated}
\title{Coerce to Calibrated}
\usage{
as_calibrated(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Coerces to calibrated output class. Generic, with flagship method
\code{\link{as_calibrated.data.frame}}. The intent is to attach
calibrated column widths for downstream table rendering.
}
\seealso{
Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated.data.frame}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl}()},
\code{\link{kbl.calibrated}()},
\code{\link{kbl.default}()},
\code{\link{kbl.gridtable}()},
\code{\link{print.gridtable}()}
}
\concept{gridtable}
\keyword{internal}
