% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{as_tablet.data.frame}
\alias{as_tablet.data.frame}
\title{Coerce data.frame to tablet}
\usage{
\method{as_tablet}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
tablet
}
\description{
Coerces data.frame to tablet.  Checks format and assigns the class.
See \code{\link{tablet.data.frame}}.
}
\seealso{
Other tablet: 
\code{\link{as_tablet}()},
\code{\link{header_rows.tablet}()},
\code{\link{header_rows}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()}
}
\concept{tablet}
