% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{as_tablet}
\alias{as_tablet}
\title{Coerce to tablet}
\usage{
as_tablet(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
tablet
}
\description{
Coerces to tablet.
Generic, with method \code{\link{as_tablet.data.frame}}.
}
\seealso{
Other tablet: 
\code{\link{as_tablet.data.frame}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()}
}
\concept{tablet}
\keyword{internal}
