% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{as_tablette.data.frame}
\alias{as_tablette.data.frame}
\title{Coerce data.frame to tablette}
\usage{
\method{as_tablette}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
tablette
}
\description{
Coerces data.frame to tablette.  Checks format and assigns the class.  See Details.
}
\details{
A tablette is a special case of data.frame with grouped rows and columns.
  * There is always one level of row groups.
  * There can be any number of column groups, including zero.
  * All columns inherit factor or character.
  * The first column is anonymous, with empty elements that represent the last non-empty value. It can be class 'latex' or 'character'.
  * Leading blank elements of first column correspond to header rows.
  * The second column is anonymous and represents group-specific property names. It is populated always and only where column 1 is not.
  * All other columns represent group-specific property values; elements before the first non-empty group value represent nested headers.
  * Header values may be repeated.
  * Header values may be empty, representing the last non-empty value to the left.
}
\seealso{
Other tablette: 
\code{\link{as_tablette}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()},
\code{\link{tablette}()}
}
\concept{tablette}
