% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{as_tablette}
\alias{as_tablette}
\title{Coerce to tablette}
\usage{
as_tablette(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
tablette
}
\description{
Coerces to tablette.
Generic, with method \code{\link{as_tablette.data.frame}}.
}
\seealso{
Other tablette: 
\code{\link{as_tablette.data.frame}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()},
\code{\link{tablette}()}
}
\concept{tablette}
\keyword{internal}
