% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.R
\name{as_xtable}
\alias{as_xtable}
\title{Coerce to Xtable}
\usage{
as_xtable(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Coerces to xtable output class.  Generic,
with method \code{\link{as_xtable.tablet}}.
}
\examples{
example(classifiers)
as_xtable(tablet(groupwise(x)))
}
\keyword{internal}
