% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{categoricals.data.frame}
\alias{categoricals.data.frame}
\title{Identify Categoricals for Data Frame}
\usage{
\method{categoricals}{data.frame}(x, ..., na.rm_fac = FALSE, exclude_fac = NULL)
}
\arguments{
\item{x}{data.frame; names must not start with '_tablet_'}

\item{...}{passed to \code{\link{classifiers}}}

\item{na.rm_fac}{whether to drop NA observations; passed to \code{\link[tidyr]{gather}} as na.rm}

\item{exclude_fac}{which factor levels to exclude; see \code{\link{factor}} (exclude)}
}
\value{
same class as x
}
\description{
Identifies categorical columns: literally, factors. Stacks factor levels and supplies value = 1.
}
\examples{
example(classifiers)
categoricals(x)
levels(categoricals(x)$level)
}
\keyword{internal}
