% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{categoricals.decorated}
\alias{categoricals.decorated}
\title{Identify Categoricals for Decorated}
\usage{
\method{categoricals}{decorated}(x, ..., na.rm_fac = FALSE, exclude_fac = NULL, all_levels = FALSE)
}
\arguments{
\item{x}{data.frame; names must not include 'name' or 'level'}

\item{...}{passed to \code{\link{classifiers}}}

\item{na.rm_fac}{whether to drop NA observations; passed to \code{\link[tidyr]{gather}} as na.rm}

\item{exclude_fac}{which factor levels to exclude; see \code{\link{factor}} (exclude)}

\item{all_levels}{whether to supply records for unobserved levels}
}
\value{
same class as x
}
\description{
Identifies categorical columns: literally, factors. Stacks factor levels and supplies value = 1.
Optionally generates placeholder records with value = 0 for unobserved factor levels.
}
\examples{
example(classifiers)
categoricals(x)
levels(categoricals(x)$level)
}
\keyword{internal}
