% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{categoricals}
\alias{categoricals}
\title{Identify Categories}
\usage{
categoricals(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\value{
see methods
}
\description{
Identifies categories.  Generic, with method \code{\link{categoricals.data.frame}}.
}
\examples{
example(classifiers)
categoricals(x)

}
\keyword{internal}
