% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{classifiers.data.frame}
\alias{classifiers.data.frame}
\title{Identify Classifiers for Data Frame}
\usage{
\method{classifiers}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
data.frame, grouped if x was (see \code{\link[dplyr]{group_by}})
}
\description{
Identifies classifiers: literally, groups. Supplies tablet_n and tablet_N: groupwise
and ungrouped counts.
}
\keyword{internal}
