% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{classifiers}
\alias{classifiers}
\title{Identify Classifiers}
\usage{
classifiers(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Identifies classifiers.  Generic, with data.frame method.
}
\examples{
library(dplyr)
x <- data.frame(grp = 1:2, a = factor(1:2), b = 1:2, c = factor(1:2, levels = 2:1))
x <- group_by(x, grp)
classifiers(x)
}
\keyword{internal}
