% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{devalued.observations}
\alias{devalued.observations}
\title{Aggregate Values for Observations
Aggregates values for observations.
Accepts a list of formulas with result name on left and
aggregating expression on the right.
Behavior is undefined if any expression does not aggregate! (I.e., length != 1).
Expressions are evaluated in an environment where values are available as 'x',
grouped count is available as 'n' and ungouped count is available as 'N'.}
\usage{
\method{devalued}{observations}(
  x,
  ...,
  fun = list(sum ~ signif(digits = 3, sum(x, na.rm = T)), pct ~ signif(digits = 3,
    sum/n * 100), ave ~ signif(digits = 3, mean(x, na.rm = T)), std ~ signif(digits = 3,
    sd(x, na.rm = T)), med ~ signif(digits = 3, median(x, na.rm = T)), min ~
    signif(digits = 3, min(x, na.rm = T)), max ~ signif(digits = 3, max(x, na.rm = T))),
  silent = TRUE
)
}
\arguments{
\item{x}{observations}

\item{...}{passed formulas with matching LHS will replace corresponding element of fun.}

\item{fun}{default aggregate functions expressed as formulas}

\item{silent}{whether to suppress warnings from evaluations of 'fun'}
}
\value{
class 'devalued'
}
\description{
Aggregate Values for Observations
Aggregates values for observations.
Accepts a list of formulas with result name on left and
aggregating expression on the right.
Behavior is undefined if any expression does not aggregate! (I.e., length != 1).
Expressions are evaluated in an environment where values are available as 'x',
grouped count is available as 'n' and ungouped count is available as 'N'.
}
\keyword{internal}
