% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{devalued}
\alias{devalued}
\title{Aggregate Values}
\usage{
devalued(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Aggregated values.  Generic, with method for 'observations'.
}
\examples{
example(classifiers)
devalued(observations(x))
}
\keyword{internal}
