% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{escape_latex.default}
\alias{escape_latex.default}
\title{Escape Latex by Default}
\usage{
\method{escape_latex}{default}(x, secondary = TRUE, ...)
}
\arguments{
\item{x}{character}

\item{secondary}{logical: whether secondary backslashes should be pre-doubled}

\item{...}{ignored}
}
\value{
escaped
}
\description{
Escapes latex markup.
Adapted with gratitude from KableExtra internals.
At kableExtra 1.3.4, sim_double_escape() only doubles
primary (leading) backslashes; the \code{secondary} argument
by default pre-doubles later backslashes to prevent
certain display errors.  Behavior may change if kableExtra changes.
}
\examples{
escape_latex('([#$\%&_{}])')
}
\seealso{
Other escape: 
\code{\link{escape_latex.escaped}()},
\code{\link{escape_latex}()}
}
\concept{escape}
\keyword{internal}
