% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{escape_latex.escaped}
\alias{escape_latex.escaped}
\title{Escape Latex for Class 'escaped'}
\usage{
\method{escape_latex}{escaped}(x, ...)
}
\arguments{
\item{x}{escaped}

\item{...}{ignored}
}
\value{
escaped
}
\description{
Returns argument unmodified.  Prevents accidental double-escaping
of the same text.
}
\examples{
identical(
escape_latex('([#$\%&_{}])'),
escape_latex(escape_latex('([#$\%&_{}])'))
)

}
\seealso{
Other escape: 
\code{\link{escape_latex.default}()},
\code{\link{escape_latex}()}
}
\concept{escape}
\keyword{internal}
