% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{groupfull}
\alias{groupfull}
\title{Calculate With Groups}
\usage{
groupfull(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\value{
see methods
}
\description{
Calculates with groups.  Generic, with method \code{\link{groupfull.data.frame}}.
It is meaningless to use this toolchain function on classes
'observations', 'devalued', or 'widgets'.
}
\keyword{internal}
