% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{groupless.data.frame}
\alias{groupless.data.frame}
\title{Calculate Widgets Without Groups}
\usage{
\method{groupless}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to observations(), devalued(), widgets()}
}
\value{
class 'groupless'
}
\description{
Calculates widgets without groups.
Removes grouping variables and groups, then executes:
data.frame -> observations -> devalued -> widgets.
}
\keyword{internal}
