% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{groupwise.data.frame}
\alias{groupwise.data.frame}
\title{Calculate Widgets With and Without Groups}
\usage{
\method{groupwise}{data.frame}(
  x,
  fun = list(sum ~ signif(digits = 3, sum(x, na.rm = T)), pct ~ signif(digits = 3,
    sum/n * 100), ave ~ signif(digits = 3, mean(x, na.rm = T)), std ~ signif(digits = 3,
    sd(x, na.rm = T)), med ~ signif(digits = 3, median(x, na.rm = T)), min ~
    signif(digits = 3, min(x, na.rm = T)), max ~ signif(digits = 3, max(x, na.rm = T))),
  fac = list(` ` ~ sum + " (" + pct + "\%" + ")"),
  num = list(`Mean (SD)` ~ ave + " (" + std + ")", `Median (range)` ~ med + " (" + min
    + ", " + max + ")"),
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{fun}{passed to groupfull() and groupless()}

\item{fac}{passed to groupfull() and groupless()}

\item{num}{passed to groupfull() and groupless()}

\item{...}{passed to groupfull() and groupless()}
}
\value{
class 'groupwise'
}
\description{
Calculates widgets with and without groups.
Supplies 'groupfull' and 'groupless' (prefixed) columns instead of 'widgets'.
Column attributes 'label' and 'title' (dominant) are substituted for column name, if present.
}
\keyword{internal}
