% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{groupwise}
\alias{groupwise}
\title{Calculate With and Without Groups}
\usage{
groupwise(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\value{
see methods
}
\description{
Calculates with and without groups.  Generic, with method \code{\link{groupwise.data.frame}}.
It is meaningless to use this toolchain function on classes
'observations', 'devalued', 'widgets', 'groupless', or 'groupfull'.
}
\examples{
example(classifiers)
groupwise(x)
}
\keyword{internal}
