% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{header_rows}
\alias{header_rows}
\title{Identify Header Rows}
\usage{
header_rows(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Identifies header rows.
Generic, with method \code{\link{header_rows.tablet}}.
}
\seealso{
Other tablet: 
\code{\link{as_tablet}()},
\code{\link{as_tablet.data.frame}()},
\code{\link{header_rows.tablet}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()}
}
\concept{tablet}
\keyword{internal}
