% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{header_rows.tablet}
\alias{header_rows.tablet}
\title{Identify Header Rows for tablet}
\usage{
\method{header_rows}{tablet}(x, ...)
}
\arguments{
\item{x}{tablet}

\item{...}{ignored}
}
\value{
integer: indices for those rows representing headers
}
\description{
Identifies header rows for tablet.
}
\seealso{
Other tablet: 
\code{\link{as_tablet.data.frame}()},
\code{\link{as_tablet}()},
\code{\link{header_rows}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()}
}
\concept{tablet}
