% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{headerlist.tablet}
\alias{headerlist.tablet}
\title{Create Header List for Tablet}
\usage{
\method{headerlist}{tablet}(x, ...)
}
\arguments{
\item{x}{tablet}

\item{...}{ignored}
}
\value{
list of named integer
}
\description{
Creates header list for tablet.
}
\keyword{internal}
