% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{headerlist.tablette}
\alias{headerlist.tablette}
\title{Create Header List for Tablet}
\usage{
\method{headerlist}{tablette}(x, ...)
}
\arguments{
\item{x}{tablet}

\item{...}{ignored}
}
\value{
list of named integer where each element describes an additional header row with names as indicated, repeated integer times
}
\description{
Creates header list for tablet.
}
\keyword{internal}
