% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{index}
\alias{index}
\title{Create Index}
\usage{
index(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\value{
see methods
}
\description{
Creates index.  Generic, with method \code{\link{index.tablette}}.
}
\examples{
example(classifiers)
index(tablette(groupwise(x)))
}
\keyword{internal}
