% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tablet.R
\name{io_csv.tablet}
\alias{io_csv.tablet}
\title{Export Tablet as CSV}
\usage{
\method{io_csv}{tablet}(x, file = "", ...)
}
\arguments{
\item{x}{tablet}

\item{file}{passed to \code{\link{io_tablet}}}

\item{...}{passed to \code{\link{io_tablet}}}
}
\value{
invisible(file)
}
\description{
Exports a tablet as comma-separated variable,
by means of \code{io_csv}.  Reminds the user
to prefer \code{io_tablet} for lossless recovery.
}
\examples{
example(io_tablet)

}
\seealso{
Other io: 
\code{\link{io_tablet}()},
\code{\link{io_tablet.character}()},
\code{\link{io_tablet.data.frame}()},
\code{\link{io_tablet.tablet}()}

Other interface: 
\code{\link{io_tablet.character}()},
\code{\link{io_tablet.data.frame}()},
\code{\link{io_tablet.tablet}()}
}
\concept{interface}
\concept{io}
\keyword{internal}
