% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tablet.R
\name{io_tablet.character}
\alias{io_tablet.character}
\title{Import Tablet}
\usage{
\method{io_tablet}{character}(x, na.strings = "NA", strip.white = FALSE, ...)
}
\arguments{
\item{x}{character file path}

\item{na.strings}{passed to \code{\link[yamlet]{io_csv.character}}}

\item{strip.white}{passed to \code{\link[yamlet]{io_csv.character}}}

\item{...}{passed to \code{\link[yamlet]{io_csv.character}}}
}
\value{
data.frame
}
\description{
Imports a tablet.
A wrapper for \code{\link[yamlet]{io_csv.character}} that also
reads associated yamlet metadata, if present, and applies it
as attributes.
}
\examples{
example(io_tablet)
}
\seealso{
Other io: 
\code{\link{io_csv.tablet}()},
\code{\link{io_tablet}()},
\code{\link{io_tablet.data.frame}()},
\code{\link{io_tablet.tablet}()}

Other interface: 
\code{\link{io_csv.tablet}()},
\code{\link{io_tablet.data.frame}()},
\code{\link{io_tablet.tablet}()}
}
\concept{interface}
\concept{io}
\keyword{internal}
