% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tablet.R
\name{io_tablet.tablet}
\alias{io_tablet.tablet}
\title{Export Tablet}
\usage{
\method{io_tablet}{tablet}(x, file = "", ...)
}
\arguments{
\item{x}{tablet}

\item{file}{passed to \code{\link[yamlet]{io_csv}}}

\item{...}{passed to \code{\link[yamlet]{io_csv}}}
}
\value{
invisible(file)
}
\description{
Exports a tablet as comma-separated variable,
as well as a yamlet version of its decorations.
Forwards to next method.
}
\examples{
example(io_tablet)
}
\seealso{
Other io: 
\code{\link{io_csv.tablet}()},
\code{\link{io_tablet}()},
\code{\link{io_tablet.character}()},
\code{\link{io_tablet.data.frame}()}

Other interface: 
\code{\link{io_csv.tablet}()},
\code{\link{io_tablet.character}()},
\code{\link{io_tablet.data.frame}()}
}
\concept{interface}
\concept{io}
\keyword{internal}
