% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{kbl.calibrated}
\alias{kbl.calibrated}
\title{Render a Calibrated Data Frame with kbl}
\usage{
\method{kbl}{calibrated}(x, ..., use_widths = TRUE)
}
\arguments{
\item{x}{A data frame with class \code{calibrated}.}

\item{...}{Additional arguments passed to \code{kbl.default}.}

\item{use_widths}{Logical scalar indicating whether column-level width
attributes should be applied with \code{\link[kableExtra]{column_spec}}. Defaults to TRUE.}
}
\value{
A kable object.
}
\description{
Render a calibrated data frame with \code{kbl} and optionally apply
column-level \code{width} attributes with \code{\link[kableExtra]{column_spec}}.
}
\seealso{
\code{\link{kbl}}, \code{\link{kbl.default}}, \code{\link[kableExtra]{column_spec}}

Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated}()},
\code{\link{as_calibrated.data.frame}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl}()},
\code{\link{kbl.default}()},
\code{\link{kbl.gridtable}()},
\code{\link{print.gridtable}()}
}
\concept{gridtable}
