% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{kbl.default}
\alias{kbl.default}
\title{Default kbl Method}
\usage{
\method{kbl}{default}(
  x,
  format,
  digits = getOption("digits"),
  row.names = NA,
  col.names = NA,
  align,
  caption = NULL,
  label = NULL,
  format.args = list(),
  escape = TRUE,
  table.attr = getOption("knitr.table.html.attr", ""),
  booktabs = FALSE,
  longtable = FALSE,
  tabular = if (longtable) "longtable" else "tabular",
  valign = if (tabular \%in\% c("tabularx", "xltabular")) "{\\\\linewidth}" else "[t]",
  position = "",
  centering = TRUE,
  vline = getOption("knitr.table.vline", if (booktabs) "" else "|"),
  toprule = getOption("knitr.table.toprule", if (booktabs) "\\\\toprule" else
    "\\\\hline"),
  bottomrule = getOption("knitr.table.bottomrule", if (booktabs) "\\\\bottomrule" else
    "\\\\hline"),
  midrule = getOption("knitr.table.midrule", if (booktabs) "\\\\midrule" else
    "\\\\hline"),
  linesep = if (booktabs) c("", "", "", "", "\\\\addlinespace") else "\\\\hline",
  caption.short = "",
  table.envir = if (!is.null(caption)) "table",
  ...
)
}
\arguments{
\item{x}{Object to render.}

\item{format, digits, row.names, col.names, align, caption, label}{Arguments
passed to \code{\link[kableExtra]{kbl}}.}

\item{format.args, escape, table.attr, booktabs, longtable, tabular, valign}{Arguments passed to \code{\link[kableExtra]{kbl}}.}

\item{position, centering, vline, toprule, bottomrule, midrule, linesep}{Arguments passed to \code{\link[kableExtra]{kbl}}.}

\item{caption.short, table.envir}{Arguments passed to
\code{\link[kableExtra]{kbl}}.}

\item{...}{Additional arguments passed to \code{\link[kableExtra]{kbl}}.}
}
\value{
A kable object.
}
\description{
Calls the existing \code{\link[kableExtra]{kbl}} implementation.
}
\seealso{
\code{\link[kableExtra]{kbl}}

Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated}()},
\code{\link{as_calibrated.data.frame}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl}()},
\code{\link{kbl.calibrated}()},
\code{\link{kbl.gridtable}()},
\code{\link{print.gridtable}()}
}
\concept{gridtable}
