% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{kbl}
\alias{kbl}
\title{Render with kbl}
\usage{
kbl(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Generic wrapper for \code{\link[kableExtra]{kbl}}. The default method is the
existing \code{kableExtra::kbl} definition.
}
\seealso{
\code{\link[kableExtra]{kbl}}

Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated}()},
\code{\link{as_calibrated.data.frame}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl.calibrated}()},
\code{\link{kbl.default}()},
\code{\link{kbl.gridtable}()},
\code{\link{print.gridtable}()}
}
\concept{gridtable}
