% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{numerics.data.frame}
\alias{numerics.data.frame}
\title{Identify Numerics for Data Frame}
\usage{
\method{numerics}{data.frame}(x, ..., na.rm_num = FALSE)
}
\arguments{
\item{x}{data.frame; names must not include 'name' or 'level'}

\item{...}{passed to \code{\link{classifiers}}}

\item{na.rm_num}{whether to drop NA observations; passed to \code{\link[tidyr]{gather}} as na.rm}
}
\value{
same class as x
}
\description{
Identifies numeric columns. Stacks values and supplies factor level 'numeric'.
}
\keyword{internal}
