% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{observations.data.frame}
\alias{observations.data.frame}
\title{Identify Observations for Data Frame}
\usage{
\method{observations}{data.frame}(x, ..., exclude_name = NULL)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to categoricals(), numerics()}

\item{exclude_name}{whether to drop NA levels of name; passed to \code{\link[tidyr]{gather}}}
}
\value{
class 'observations' arranged by groups (presented first) and including: names, levels, N, n, value (prefixed)
}
\description{
Identifies observations. Stacks values by level, supplying defaults as necessary.
}
\keyword{internal}
