% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{observations}
\alias{observations}
\title{Identify Observations}
\usage{
observations(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Identifies observations.  Generic, with data.frame method.
}
\examples{
example(classifiers)
observations(x)
}
\keyword{internal}
