% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtable.R
\name{print.gridtable}
\alias{print.gridtable}
\title{Print a gridtable}
\usage{
\method{print}{gridtable}(x, ...)
}
\arguments{
\item{x}{Character scalar with class \code{gridtable}.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a gridtable as plain grid-table text.
}
\seealso{
Other gridtable: 
\code{\link{as.data.frame.gridtable}()},
\code{\link{as_calibrated}()},
\code{\link{as_calibrated.data.frame}()},
\code{\link{as_gridtable}()},
\code{\link{as_gridtable.character}()},
\code{\link{as_gridtable.data.frame}()},
\code{\link{kbl}()},
\code{\link{kbl.calibrated}()},
\code{\link{kbl.default}()},
\code{\link{kbl.gridtable}()}
}
\concept{gridtable}
