% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{recap.knitr_kable}
\alias{recap.knitr_kable}
\title{Recap knitr_kable.}
\usage{
\method{recap}{knitr_kable}(x, cols = NULL, pos = "c", ...)
}
\arguments{
\item{x}{object of dispatch}

\item{cols}{number of columns to span; guesses \code{ncol(x)} by default}

\item{pos}{position of text: 'l','c' (default), or 'r'}

\item{...}{ignored}
}
\description{
Recaps knitr_kable.
Specifically, it replaces the first non-tabled caption
with multicolumn text.  The intent is to
prevent repeat bookmarks when generating pdf.
}
\seealso{
Other recap: 
\code{\link{recap}()}
}
\concept{recap}
