% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{recap}
\alias{recap}
\title{Recap Something}
\usage{
recap(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Recaps something.
Generic, with method \code{\link{recap.knitr_kable}}.
}
\seealso{
Other recap: 
\code{\link{recap.knitr_kable}()}
}
\concept{recap}
\keyword{internal}
