% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{splice}
\alias{splice}
\title{Splice Some Things Together}
\usage{
splice(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Spices some things together. Generic, with tablet-oriented method \code{\link{splice.data.frame}}.
}
\examples{
example(splice.data.frame)
}
\keyword{internal}
