% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{tablet.decorated}
\alias{tablet.decorated}
\title{Generate a Tablet for Decorated}
\usage{
\method{tablet}{decorated}(x, ..., all_levels = FALSE)
}
\arguments{
\item{x}{decorated}

\item{...}{passed arguments}

\item{all_levels}{whether to supply records for unobserved levels}
}
\value{
tablet
}
\description{
Generates a tablet for decorated data.frame.
Simply passes argument \code{all_levels} to
\code{\link{tablet.data.frame}}, where it is
intercepted downstream by \code{\link{categoricals.decorated}}.
Can be used to force the display of levels not present in the data.
}
