% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{tablet.groupwise}
\alias{tablet.groupwise}
\title{Generate a Group-wise Tablet}
\usage{
\method{tablet}{groupwise}(x, ..., all = "All", lab = list(lab ~ name + "\\n(N = " + n + ")"))
}
\arguments{
\item{x}{groupwise}

\item{...}{ignored}

\item{all}{a column name for ungrouped statistics}

\item{lab}{a list of formulas to generate column labels; \\n is translated as <br> in html context}
}
\value{
tablet, having columns 'name', 'level', 'stat', 'sort' (prefixed with '_tablet'); group columns (if any), and 'all' column; sorted by 'name' and 'sort'.
}
\description{
Generates a group-wise tablet. Calculates statistics
for all factors and numerics, with and without grouping
variables.  Column names represent finest level of
grouping, distinguished by attribute 'nest' (the values of
higher other groups). Column names include 'all' column
for same statistics without groups. Result columns
have corresponding attribute 'n'. 'lab' supplies a label attribute
for each column where the RHS succeeds: by default appending 'n' to result column names.
Column attributes 'label' and 'title' (dominant) are substituted for column name, if present.
}
\keyword{internal}
