% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{tablet}
\alias{tablet}
\title{Generate a Tablet}
\usage{
tablet(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Generates a tablet.  Generic. See \code{\link{tablet.data.frame}} and \code{\link{as_kable.tablet}}.
}
\keyword{internal}
