% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{tablet.tablette}
\alias{tablet.tablette}
\title{Convert tablette to tablet}
\usage{
\method{tablet}{tablette}(x, ...)
}
\arguments{
\item{x}{tablette}

\item{...}{passed arguments}
}
\value{
tablet data.frame
}
\description{
Converts tablette to tablet.
Intends to be the approximate inverse of \code{\link{tablette.tablet}}. See also \code{\link{as_tablette.data.frame}}.
}
\examples{
example(tablette.tablet)
}
\seealso{
Other tablette: 
\code{\link{as_tablette.data.frame}()},
\code{\link{as_tablette}()},
\code{\link{tablette.tablet}()},
\code{\link{tablette}()}
}
\concept{tablet}
\concept{tablette}
