% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{tablette}
\alias{tablette}
\title{Convert to tablette}
\usage{
tablette(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Converts to 'tablette'.  Generic,
with method \code{\link{tablette.tablet}}.
}
\seealso{
Other tablette: 
\code{\link{as_tablette.data.frame}()},
\code{\link{as_tablette}()},
\code{\link{tablet.tablette}()},
\code{\link{tablette.tablet}()}
}
\concept{tablette}
\keyword{internal}
