% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{widgets.devalued}
\alias{widgets.devalued}
\title{Calculate Widgets for Devalued}
\usage{
\method{widgets}{devalued}(
  x,
  fac = list(` ` ~ sum + " (" + pct + "\%" + ")"),
  num = list(`Mean (SD)` ~ ave + " (" + std + ")", `Median (range)` ~ med + " (" + min
    + ", " + max + ")"),
  ...
)
}
\arguments{
\item{x}{devalued}

\item{fac}{a list of formulas to generate widgets for factors}

\item{num}{a list of formulas to generate widgets for numerics}

\item{...}{formulas with matching LHS replace defaults}
}
\value{
class 'widgets', with new columns 'stat' and 'widget' (prefixed), arranged by groups
}
\description{
Calculates widgets for class 'devalued'.
In this context, a widget is a text fragment
that formats one or more aggregate statistics,
all of which must have been present as LHS of
some element of 'fun' in the call to devalued().
Furthermore, LHS must be unique across 'fac' and
'num' for expected behavior.
}
\details{
In the context of this call, "+" is redefined
to allow concatenation of text.
Evaluation proceeds left to right as usual.
}
\keyword{internal}
