% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{categoricals}
\alias{categoricals}
\title{Identify Categories}
\usage{
categoricals(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Identifies categories.  Generic, with data.frame method.
}
\examples{
example(classifiers)
categoricals(x)

}
\keyword{internal}
