% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape Latex.}
\usage{
escape_latex(x, ...)
}
\arguments{
\item{x}{presumably character}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Escaples latex.  Generic, with method \code{link{escape_latex.default}}.
}
\examples{
example(escape_latex.default)
}
\seealso{
Other escape: 
\code{\link{escape_latex.default}()},
\code{\link{escape_latex.escaped}()}
}
\concept{escape}
\keyword{internall}
