% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape Latex}
\usage{
escape_latex(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Escapes latex markup.
Adapted with gratitude from KableExtra internals.
}
\examples{
escape_latex('([#$\%&_{}])')
}
\keyword{internal}
