% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{groupless}
\alias{groupless}
\title{Calculate Without Groups}
\usage{
groupless(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Calculates without groups.  Generic, with method for data.frame.
It is meaningless to use this toolchain function on classes
'observations', 'devalued', or 'widgets'.
}
\keyword{internal}
