% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{headerlist}
\alias{headerlist}
\title{Create Header List.}
\usage{
headerlist(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Creates header list.  Generic, with method for tablet.
}
\examples{
example(classifiers)
headerlist(tablet(groupwise(x)))
}
\keyword{internal}
