% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{index.tablet}
\alias{index.tablet}
\title{Create Index for Tablet}
\usage{
\method{index}{tablet}(x, ...)
}
\arguments{
\item{x}{tablet}

\item{...}{ignored}
}
\value{
named integer giving row groupings with names as indicated, repeated integer times
}
\description{
Creates index for tablet.
}
\keyword{internal}
