% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa.R
\name{mesa}
\alias{mesa}
\title{Drag-and-drop Descriptive Statistics}
\usage{
mesa()
}
\value{
used for side effects: launches shiny application
}
\description{
Generate a table of descriptive statistics
by selecting columns from a file,
Currently supported formats include *.xpt, *.sas7bdat, and *.csv.
XPT formats and CSV formats are supported.
Launch the application using \code{mesa()} and use the interface
to select a data file. Then classify the
columns of interest to generate the corresponding
displays.
}
\details{
Currently,

* xpt files are read using the defaults for \code{\link[foreign]{read.xport}},

* sas7bdat files are read using the defaults for \code{\link[haven]{read_sas}}, and

* csv files are read using the defaults for \code{\link[csv]{as.csv}}.

If a file in the same directory has a corresponding base name but a \code{.yaml}
extension, it is treated as metadata and an attempt is made to apply it
to the internal version of the data. This file will not be over-written,
but it WILL be constructed if missing.  You can hand-edit it to supply
metadata.  See \code{?yamlet}.

The \code{\link[datasets]{mtcars}} datasets in the 'examples' volume is from \pkg{datasets}.
}
