% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{numerics}
\alias{numerics}
\title{Identify Numerics}
\usage{
numerics(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\value{
see methods
}
\description{
Identifies numerics.  Generic, with method \code{\link{numerics.data.frame}}.
}
\examples{
example(classifiers)
numerics(x)
}
\keyword{internal}
