% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{observations.data.frame}
\alias{observations.data.frame}
\title{Identify Observations for Data Frame}
\usage{
\method{observations}{data.frame}(x, ..., exclude_name = NULL)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to categoricals(), numerics()}

\item{exclude_name}{whether to drop NA levels of name; passed to \code{\link[tidyr]{gather}}}
}
\value{
class 'observations' arranged by groups (presented first):
\item{_tablet_N}{number of records}
\item{_tablet_n}{number of records in group}
\item{_tablet_name}{observation identifier}
\item{_tablet_level}{factor level (or special value 'numeric' for numerics)}
\item{_tablet_value}{numeric value (or special value 1 for factors)}
}
\description{
Identifies observations. Stacks values by level, supplying defaults as necessary.
}
\keyword{internal}
