% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablette.R
\name{tablette.tablet}
\alias{tablette.tablet}
\title{Convert to tablette from tablet}
\usage{
\method{tablette}{tablet}(x, ...)
}
\arguments{
\item{x}{tablette}

\item{...}{passed arguments}
}
\value{
tablet data.frame
}
\description{
Converts to tablette from tablet.
Intends to be the inverse of \code{\link{tablet.tablette}}.
}
\examples{
example(tablet.tablette)
}
\seealso{
Other tablette: 
\code{\link{tablette}()}

Other tablet: 
\code{\link{as_tablet.data.frame}()},
\code{\link{as_tablet}()},
\code{\link{tablet.tablette}()}
}
\concept{tablet}
\concept{tablette}
\keyword{internal}
