% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{widgets}
\alias{widgets}
\title{Calculate Widgets}
\usage{
widgets(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed}
}
\description{
Calculates widgets.  Generic, with method for devalued.
}
\examples{
example(classifiers)
widgets(devalued(observations(x)))
}
\keyword{internal}
